Revert `erc20_openzeppelin.cairo` to compat `cairo-lang==0.9.1`.
Basically, we previously copied in a version of the above file from the
`OpenZeppelin/cairo-contracts` repo that is incompatible with
`horus-compile`, since it uses the new `cairo-lang>=0.10.0` syntax for
stuff. This commit swaps in a version from an older revision of
`cairo-contracts`, and `horus-compile` can now at least parse it
correctly.