Refactor logger `Impl` type from `StateT` to `State`.
This commit removes use of monad transformers from `Logger/Runner.hs`,
as they are no longer needed. In particular, we change the constructor:
```
ImplT (StateT (Seq Message) m a)
```
to
```
ImplL (State (Seq Message) a)
```.

We also add an `Either Text` context around `runImpl` so it is
compatible with `GlobalL`.

* Format all source files with `fourmolu`.